BE-FR-019_Feature Histories Modules (Medium) - Weapon,events, and lilin.g rajin

+++ app/Http/Controllers/EventsController.php
@@ -9,6 +9,7 @@ use App\Helpers\Validation;
 use App\Helpers\Generator;
 
 use App\Models\Events;
+use App\Models\Histories;
 
 class EventsController extends Controller
 {
@@ -17,6 +18,7 @@ class EventsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+    /
     public function createEvent(Request $request)
     {
         try {
@@ -33,22 +35,49 @@ class EventsController extends Controller
                 $check = Events::selectRaw('1')->where('event', $request->event)->first();
                 
                 if($check == null){
-                    $uuid = Generator::getUUID();
-                    Events::create([
-                        'id' => $uuid,
-                        'event' => $request->event,
-                        'date_start' => $request->date_start,
-                        'date_end' => $request->date_end,
-                        'created_at' => date('Y-m-d H:i:s'),
-                        'created_by' => "1",
-                        'updated_at' => null,
-                        'updated_by' => null,
-                    ]);
-            
-                    return response()->json([
-                        "message" => Generator::getMessageTemplate("api_create", "event", $request->event),
-                        "status" => 'success'
-                    ], Response::HTTP_OK);
+                    $msg = Generator::getMessageTemplate("api_create", "event", $request->name);
+                    $data = new Request();
+                    $obj = [
+                        'type' => "event",
+                        'body' => $msg
+                    ];
+                    $data->merge($obj);
+
+                    $validatorHistory = Validation::getValidateHistory($data);
+
+                    if ($validatorHistory->fails()) {
+                        $errors = $validatorHistory->messages();
+
+                        return response()->json([
+                            'status' => 'failed',
+                            'result' => $errors,
+                        ], Response::HTTP_UNPROCESSABLE_ENTITY);
+                    } else {   
+                        $uuid = Generator::getUUID();
+                        Events::create([
+                            'id' => $uuid,
+                            'event' => $request->event,
+                            'date_start' => $request->date_start,
+                            'date_end' => $request->date_end,
+                            'created_at' => date('Y-m-d H:i:s'),
+                            'created_by' => "1",
+                            'updated_at' => null,
+                            'updated_by' => null,
+                        ]);
+
+                        Histories::create([
+                            'id' => Generator::getUUID(),
+                            'history_type' => $data->type, 
+                            'body' => $data->body,
+                            'created_at' => date("Y-m-d H:i:s"),
+                            'created_by' => '1' // for now
+                        ]);
+                
+                        return response()->json([
+                            "message" => Generator::getMessageTemplate("api_create", "event", $request->event),
+                            "status" => 'success'
+                        ], Response::HTTP_OK);
+                    }
                 }else{
                     return response()->json([
                         "message" => "Data is already exist", 
@@ -95,18 +124,45 @@ class EventsController extends Controller
                     "status" => 'error'
                 ], Response::HTTP_UNPROCESSABLE_ENTITY);
             } else {
-                Events::where('id', $id)->update([
-                    'event' => $request->event,
-                    'date_start' => $request->date_start,
-                    'date_end' => $request->date_end,
-                    'updated_at' => date('Y-m-d H:i:s'),
-                    'updated_by' => null,
-                ]);
-        
-                return response()->json([
-                    "message" => Generator::getMessageTemplate("api_update", "airplane", $request->event),
-                    "status" => 'success'
-                ], Response::HTTP_OK);
+                $msg = Generator::getMessageTemplate("api_update", "event", $request->name);
+                $data = new Request();
+                $obj = [
+                    'type' => "event",
+                    'body' => $msg
+                ];
+                $data->merge($obj);
+
+                $validatorHistory = Validation::getValidateHistory($data);
+
+                if ($validatorHistory->fails()) {
+                    $errors = $validatorHistory->messages();
+
+                    return response()->json([
+                        'status' => 'failed',
+                        'result' => $errors,
+                    ], Response::HTTP_UNPROCESSABLE_ENTITY);
+                } else {   
+                    Events::where('id', $id)->update([
+                        'event' => $request->event,
+                        'date_start' => $request->date_start,
+                        'date_end' => $request->date_end,
+                        'updated_at' => date('Y-m-d H:i:s'),
+                        'updated_by' => null,
+                    ]);
+
+                    Histories::create([
+                        'id' => Generator::getUUID(),
+                        'history_type' => $data->type, 
+                        'body' => $data->body,
+                        'created_at' => date("Y-m-d H:i:s"),
+                        'created_by' => '1' // for now
+                    ]);
+            
+                    return response()->json([
+                        "message" => $msg,
+                        "status" => 'success'
+                    ], Response::HTTP_OK);
+                }
             }
         } catch(\Exception $e) {
             return response()->json([
@@ -121,13 +177,41 @@ class EventsController extends Controller
             $evt = Events::select('event')
                 ->where('id', $id)
                 ->first();
-                
-            Events::destroy($id);
 
-            return response()->json([
-                'message' => Generator::getMessageTemplate("api_delete", "event",$evt->event), 
-                'status' => 'success'
-            ], Response::HTTP_OK);
+            $msg = Generator::getMessageTemplate("api_delete", "event",$evt->event);
+            $data = new Request();
+            $obj = [
+                'type' => "event",
+                'body' => $msg
+            ];
+            $data->merge($obj);
+
+            $validatorHistory = Validation::getValidateHistory($data);
+
+            if ($validatorHistory->fails()) {
+                $errors = $validatorHistory->messages();
+
+
+                Histories::create([
+                    'id' => Generator::getUUID(),
+                    'history_type' => $data->type, 
+                    'body' => $data->body,
+                    'created_at' => date("Y-m-d H:i:s"),
+                    'created_by' => '1' // for now
+                ]);
+                
+                return response()->json([
+                    'status' => 'failed',
+                    'result' => $errors,
+                ], Response::HTTP_UNPROCESSABLE_ENTITY);
+            } else {   
+                Events::destroy($id);
+
+                return response()->json([
+                    'message' => $msg, 
+                    'status' => 'success'
+                ], Response::HTTP_OK);
+            }
         } catch(\Exception $e) {
             return response()->json([
                 'status' => 'error',
